+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v3.1.2
 * Fixed: Find path when edge labels are deflated
 * Fixed: Accumulo stream property value in table data length of reference
+* Fixed: Query will now return hidden elements/vertices/edges if FetchHint.INCLUDE_HIDDEN is passed
+* Added: Query methods elementIds/vertexIds/edgeIds are overloaded to accept IdFetchHint, which makes it possible to include ids for hidden elements
 
 # v3.1.1
 * Added: Graph.getExtendedData to get a single extended data row